backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
qdrant-client==1.8.2
Pillow==10.3.0
requests==2.31.0
torch==2.2.2
torchvision==0.17.2
transformers==4.39.3
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
python-dotenv==1.0.1
boto3==1.34.84
numpy<2.0.0
httpx==0.27.0
tenacity==8.3.0
